[add]重複チェック用のRepositoryメソッドを追加

+++ src/main/java/org/example/persistence/repository/UserGroupBelongingRepository.java
@@ -1,6 +1,7 @@
 package org.example.persistence.repository;
 
 import org.example.persistence.entity.UserGroupBelonging;
+import org.springframework.data.r2dbc.repository.Query;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
@@ -19,4 +20,12 @@ public interface UserGroupBelongingRepository extends
   Mono<UserGroupBelonging> save(UserGroupBelonging userGroupBelonging);
 
   Mono<Void> deleteById(Long id);
+
+  @Query("""
+      SELECT * FROM rbac_user_group_belongings
+      WHERE namespace_id = :namespaceId
+        AND user_id = :userId
+        AND user_group_id = :userGroupId;
+      """)
+  Mono<UserGroupBelonging> findDuplicated(Long namespaceId, Long userId, Long userGroupId);
 }
